requirements.txt
anyio==3.7.1
base58==2.1.1
certifi==2025.7.14
Deprecated==1.2.18
ECPy==1.2.5
fastapi==0.104.1
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.10
pycryptodome==3.23.0
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.1
supabase==2.3.4
types-Deprecated==1.2.15.20250304
typing_extensions==4.14.1
uvicorn==0.24.0
websockets==12.0
wrapt==1.17.2
xrpl-py==4.2.0
python-jose[cryptography]==3.3.0
